cleanup , split model into separate file
